refactor proposal memo routes
