+++ src/app/routers/proposal_memo_evidence.py
@@ -1,0 +1,85 @@
+from fastapi import APIRouter, Query
+
+from app.contracts.proposals import (
+    ProposalMemoLineageEnvelopeResponse,
+    ProposalMemoProjectionEnvelopeResponse,
+    ProposalMemoReplayEvidenceEnvelopeResponse,
+)
+from app.middleware.correlation import correlation_id_var
+from app.routers.proposal_memo_common import PROPOSAL_ID_PATH, VERSION_NO_PATH
+from app.services.advisory_service_provider import proposal_service
+
+router = APIRouter(prefix="/api/v1/proposals", tags=["proposals"])
+
+
+@router.get(
+    "/{proposal_id}/versions/{version_no}/memo/projection",
+    response_model=ProposalMemoProjectionEnvelopeResponse,
+    summary="Get Proposal Memo Projection",
+    description=(
+        "Returns an audience-specific memo projection from lotus-advise for advisor, compliance, "
+        "operations, or client-draft review. Gateway forwards the requested audience and does not "
+        "redact, rank, or construct memo content locally."
+    ),
+)
+async def get_proposal_memo_projection(
+    proposal_id: str = PROPOSAL_ID_PATH,
+    version_no: int = VERSION_NO_PATH,
+    audience: str | None = Query(
+        default=None,
+        description="Optional lotus-advise memo projection audience.",
+        examples=["COMPLIANCE"],
+    ),
+) -> ProposalMemoProjectionEnvelopeResponse:
+    service = proposal_service()
+    correlation_id = correlation_id_var.get()
+    return await service.get_proposal_memo_projection(
+        proposal_id=proposal_id,
+        version_no=version_no,
+        audience=audience,
+        correlation_id=correlation_id,
+    )
+
+
+@router.get(
+    "/{proposal_id}/memos/lineage",
+    response_model=ProposalMemoLineageEnvelopeResponse,
+    summary="Get Proposal Memo Lineage",
+    description=(
+        "Returns proposal memo lineage from lotus-advise, including memo hashes, review posture, "
+        "report-package posture, archive refs, replay evidence, and AI commentary posture without "
+        "gateway-side recomputation."
+    ),
+)
+async def get_proposal_memo_lineage(
+    proposal_id: str = PROPOSAL_ID_PATH,
+) -> ProposalMemoLineageEnvelopeResponse:
+    service = proposal_service()
+    correlation_id = correlation_id_var.get()
+    return await service.get_proposal_memo_lineage(
+        proposal_id=proposal_id,
+        correlation_id=correlation_id,
+    )
+
+
+@router.get(
+    "/{proposal_id}/versions/{version_no}/memo/replay-evidence",
+    response_model=ProposalMemoReplayEvidenceEnvelopeResponse,
+    summary="Get Proposal Memo Replay Evidence",
+    description=(
+        "Returns memo replay evidence from lotus-advise so operations can inspect source hashes, "
+        "audit events, supportability, and blocked client-ready posture without local Gateway "
+        "interpretation."
+    ),
+)
+async def get_proposal_memo_replay_evidence(
+    proposal_id: str = PROPOSAL_ID_PATH,
+    version_no: int = VERSION_NO_PATH,
+) -> ProposalMemoReplayEvidenceEnvelopeResponse:
+    service = proposal_service()
+    correlation_id = correlation_id_var.get()
+    return await service.get_proposal_memo_replay_evidence(
+        proposal_id=proposal_id,
+        version_no=version_no,
+        correlation_id=correlation_id,
+    )
